Personnalised error403 & error404 pages + Back access control Gestion Filiere

+++ src/Controller/FiliereController.php
@@ -9,11 +9,13 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
+
 
 #[Route('admin/filiere')]
 class FiliereController extends AbstractController
 {
-    #[Route('/', name: 'app_filiere_index', methods: ['GET'])]
+    #[Route('/', name: 'app_filiere_index', methods: ['GET']), IsGranted('ROLE_ADMIN')]
     public function index(FiliereRepository $filiereRepository): Response
     {
         $user = $this->getUser();
@@ -27,9 +29,14 @@ class FiliereController extends AbstractController
         ]);
     }
 
-    #[Route('/new', name: 'app_filiere_new', methods: ['GET', 'POST'])]
+    #[Route('/new', name: 'app_filiere_new', methods: ['GET', 'POST']), IsGranted('ROLE_ADMIN')]
     public function new(Request $request, FiliereRepository $filiereRepository): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         $filiere = new Filiere();
         $form = $this->createForm(FiliereType::class, $filiere);
         $form->handleRequest($request);
@@ -46,9 +53,14 @@ class FiliereController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_filiere_show', methods: ['GET'])]
+    #[Route('/{id}', name: 'app_filiere_show', methods: ['GET']), IsGranted('ROLE_ADMIN')]
     public function show(Filiere $filiere): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         $entityManager = $this->getDoctrine()->getManager();
         $filiere = $entityManager->getRepository(Filiere::class)->find($filiere);
         $users = $filiere->getUsers();
@@ -59,9 +71,14 @@ class FiliereController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'app_filiere_edit', methods: ['GET', 'POST'])]
+    #[Route('/{id}/edit', name: 'app_filiere_edit', methods: ['GET', 'POST']), IsGranted('ROLE_ADMIN')]
     public function edit(Request $request, Filiere $filiere, FiliereRepository $filiereRepository): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         $form = $this->createForm(FiliereType::class, $filiere);
         $form->handleRequest($request);
 
@@ -77,9 +94,14 @@ class FiliereController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'app_filiere_delete', methods: ['POST'])]
+    #[Route('/{id}', name: 'app_filiere_delete', methods: ['POST']), IsGranted('ROLE_ADMIN')]
     public function delete(Request $request, Filiere $filiere, FiliereRepository $filiereRepository): Response
     {
+        $user = $this->getUser();
+
+        if (!$user) {
+            return $this->redirectToRoute('app_login');
+        }
         if ($this->isCsrfTokenValid('delete'.$filiere->getId(), $request->request->get('_token'))) {
             $filiereRepository->remove($filiere, true);
         }
